assets/08-vorlesung/src/requirements.txt
azure_common==1.1.21
paho_mqtt==1.4.0
